Uploaded game from previous assignment
The game uploaded here is the one from our previous assignment -
basically it adheres to all the functionality required in the game. When
the player enters the game a choice is given to start or exit. On
pressing start the user is prompted to enter their username before the
game begins. When the game begins enemies are spawned and the player
must destroy them to advance to the next level. By pressing space the
player shoots and by pressing e the player activates the shield. When
the player is hit by their lasers his health decreases and when they are
hit by the player's laser the aliens die. The player's stats are
displayed at the top left of the game. A power up is dropped randomly
that replenishes shield/health